+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "app-soccer",
+  "name": "soccer-premierleague",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
